Refactored the main package
